# [Models] Align State Entity with Database Schema