--- a/src/main/java/com/iws_manager/iws_manager_api/models/State.java
+++ b/src/main/java/com/iws_manager/iws_manager_api/models/State.java
@@ -18,11 +18,12 @@ import lombok.*;
 @EqualsAndHashCode(callSuper = true)
 @EntityListeners(AuditingEntityListener.class)
 @Table(name = "state")
+@AttributeOverride(name = "id", column = @Column(name = "stateid"))
 public class State extends BaseEntity{
 
     /**
      * Official name of the state (e.g., "Bavaria", "Berlin", "Hamburg").
      */
-    @Column(name = "name", nullable = false)
+    @Column(name = "statename", nullable = true, length = 255)
     private String name;
 }